MAINT: move module from operator to discr
